*** empty log message ***

+++ rose/modules/Rose-DB-Object/t/db-object-cached.t
@@ -2,7 +2,7 @@
 
 use strict;
 
-use Test::More tests => 337;
+use Test::More tests => 407;
 
 BEGIN 
 {
@@ -10,7 +10,7 @@ BEGIN
   use_ok('Rose::DB::Object::Cached');
 }
 
-our($PG_HAS_CHKPASS, $HAVE_PG, $HAVE_MYSQL, $HAVE_INFORMIX);
+our($PG_HAS_CHKPASS, $HAVE_PG, $HAVE_MYSQL, $HAVE_INFORMIX, $HAVE_SQLITE);
 
 #
 # Generic
@@ -608,6 +608,191 @@ SKIP: foreach my $db_type (qw(informix))
   ok($MyInformixObject::Objects_By_Key_Loaded{'name'}{'John'} != $loaded, "cache_expires_in uk 3 - $db_type");
 }
 
+#
+# SQLite
+#
+
+SKIP: foreach my $db_type (qw(sqlite))
+{
+  skip("SQLite tests", 70)  unless($HAVE_SQLITE);
+
+  Rose::DB->default_type($db_type);
+
+  my $of = MySQLiteObject->new(name => 'John', id => 99);
+
+  ok(ref $of && $of->isa('MySQLiteObject'), "cached new() 1 - $db_type");
+
+  ok($of->save, "save() 1 - $db_type");
+
+  my $of2 = MySQLiteObject->new(id => $of->id);
+
+  ok(ref $of2 && $of2->isa('MySQLiteObject'), "cached new() 2 - $db_type");
+
+  ok($of2->load, "cached load() - $db_type");
+
+  is($of2->name, $of->name, "load() verify 1 - $db_type");
+
+  my $of3 = MySQLiteObject->new(id => $of2->id);
+
+  ok(ref $of3 && $of3->isa('MySQLiteObject'), "cached new() 3 - $db_type");
+
+  ok($of3->load, "cached load() - $db_type");
+
+  is($of3->name, $of2->name, "cached load() verify 2 - $db_type");
+
+  is($of3, $of2, "load() verify cached 1 - $db_type");
+  is($of2, $of, "load() verify cached 2 - $db_type");
+
+  my $ouk = MySQLiteObject->new(name => $of->name);
+
+  ok($ouk->load, "cached load() unique key - $db_type");
+  is($ouk, $of, "load() verify cached unique key 1 - $db_type");
+  is($ouk, $of2, "load() verify cached unique key 2 - $db_type");
+  is($ouk, $of3, "load() verify cached unique key 3 - $db_type");
+
+  is(keys %MySQLiteObject::Objects_By_Id, 1, "cache check 1 - $db_type");
+
+  ok($of->forget, "forget() - $db_type");
+
+  is(keys %MySQLiteObject::Objects_By_Id, 0, "cache check 2 - $db_type");
+
+  # Standard tests
+
+  my $o = MySQLiteObject->new(name => 'John x', id => 1);
+
+  ok(ref $o && $o->isa('MySQLiteObject'), "new() 1 - $db_type");
+
+  $o->flag2('true');
+  $o->date_created('now');
+  $o->last_modified($o->date_created);
+  $o->save_col(7);
+
+  ok($o->save, "save() 1 - $db_type");
+  ok($o->load, "load() 1 - $db_type");
+
+  my $o2 = MySQLiteObject->new(id => $o->id);
+
+  ok(ref $o2 && $o2->isa('MySQLiteObject'), "new() 2 - $db_type");
+
+  is($o2->bits->to_Bin, '00101', "bits() (bitfield default value) - $db_type");
+
+  ok($o2->load, "load() 2 - $db_type");
+  ok(!$o2->not_found, "not_found() 1 - $db_type");
+
+  is($o2->name, $o->name, "load() verify 1 - $db_type");
+  is($o2->date_created, $o->date_created, "load() verify 2 - $db_type");
+  is($o2->last_modified, $o->last_modified, "load() verify 3 - $db_type");
+  is($o2->status, 'active', "load() verify 4 (default value) - $db_type");
+  is($o2->flag, 1, "load() verify 5 (default boolean value) - $db_type");
+  is($o2->flag2, 1, "load() verify 6 (boolean value) - $db_type");
+  is($o2->save_col, 7, "load() verify 7 (aliased column) - $db_type");
+  is($o2->start->ymd, '1980-12-24', "load() verify 8 (date value) - $db_type");
+
+  is($o2->bits->to_Bin, '00101', "load() verify 9 (bitfield value) - $db_type");
+
+  $o2->name('John 2');
+  $o2->start('5/24/2001');
+
+  sleep(1); # keep the last modified dates from being the same
+
+  $o2->last_modified('now');
+
+  ok($o2->save, "save() 2 - $db_type");
+  ok($o2->load, "load() 3 - $db_type");
+
+  is($o2->date_created, $o->date_created, "save() verify 1 - $db_type");
+  ok($o2->last_modified eq $o->last_modified, "save() verify 2 - $db_type");
+  is($o2->start->ymd, '2001-05-24', "save() verify 3 (date value) - $db_type");
+
+  my $o3 = MySQLiteObject->new();
+
+  my $db = $o3->db or die $o3->error;
+
+  ok(ref $db && $db->isa('Rose::DB'), "db() - $db_type");
+
+  is($db->dbh, $o3->dbh, "dbh() - $db_type");
+
+  my $o4 = MySQLiteObject->new(id => 999);
+  ok(!$o4->load(speculative => 1), "load() nonexistent - $db_type");
+  ok($o4->not_found, "not_found() 2 - $db_type");
+
+  ok($o->load, "load() 4 - $db_type");
+
+  my $o5 = MySQLiteObject->new(id => $o->id);
+
+  ok($o5->load, "load() 5 - $db_type");
+
+  $o5->nums([ 4, 5, 6 ]);
+  ok($o5->save, "save() 4 - $db_type");
+  ok($o->load, "load() 6 - $db_type");
+
+  is($o5->nums->[0], 4, "load() verify 10 (array value) - $db_type");
+  is($o5->nums->[1], 5, "load() verify 11 (array value) - $db_type");
+  is($o5->nums->[2], 6, "load() verify 12 (array value) - $db_type");
+
+  my @a = $o5->nums;
+
+  is($a[0], 4, "load() verify 13 (array value) - $db_type");
+  is($a[1], 5, "load() verify 14 (array value) - $db_type");
+  is($a[2], 6, "load() verify 15 (array value) - $db_type");
+  is(@a, 3, "load() verify 6 (array value) - $db_type");
+
+  ok($o->delete, "delete() - $db_type");
+
+  eval { $o->meta->alias_column(nonesuch => 'foo') };
+  ok($@, 'alias_column() nonesuch');
+
+  $o2->forget;
+
+  $o = MySQLiteObject->new(name => 'John');
+  ok($o->load, "load() forget 1 - $db_type");
+
+  $o->forget;
+
+  $o2 = MySQLiteObject->new(name => 'John');
+  ok($o2->load, "load() forget 2 - $db_type");
+
+  ok($o ne $o2, "load() forget 3 - $db_type");
+
+  $o->meta->clear_object_cache;
+
+  FORGET_ALL_SQLITE:
+  {
+    no warnings;
+    is(scalar keys %MySQLiteObject::Objects_By_Id, 0, "clear_object_cache() 1 - $db_type");
+    is(scalar keys %MySQLiteObject::Objects_By_Key, 0, "clear_object_cache() 2 - $db_type");
+    is(scalar keys %MySQLiteObject::Objects_Keys, 0, "clear_object_cache() 3 - $db_type");
+  }
+
+  # Cache expiration with primary key
+  MySQLiteObject->meta->cached_objects_expire_in('5 seconds');
+  $o = MySQLiteObject->new(id => 99);
+  $o->load or die $o->error;
+
+  my $loaded = $MySQLiteObject::Objects_By_Id_Loaded{99};
+
+  is($MySQLiteObject::Objects_By_Id_Loaded{99}, $loaded, "cache_expires_in pk 1 - $db_type");
+  $o->load or die $o->error;
+  is($MySQLiteObject::Objects_By_Id_Loaded{99}, $loaded, "cache_expires_in pk 2 - $db_type");
+  sleep(5);
+  $o->load or die $o->error;
+  ok($MySQLiteObject::Objects_By_Id_Loaded{99} != $loaded, "cache_expires_in pk 3 - $db_type");
+
+  # Cache expiration with unique key
+  MySQLiteObject->meta->cached_objects_expire_in('5 seconds');
+  $o = MySQLiteObject->new(name => 'John');
+  $o->load or die $o->error;
+
+  $loaded = $MySQLiteObject::Objects_By_Key_Loaded{'name'}{'John'};
+
+  is($MySQLiteObject::Objects_By_Key_Loaded{'name'}{'John'}, $loaded, "cache_expires_in uk 1 - $db_type");
+  $o->load or die $o->error;
+  is($MySQLiteObject::Objects_By_Key_Loaded{'name'}{'John'}, $loaded, "cache_expires_in uk 2 - $db_type");
+  sleep(5);
+  $o->load or die $o->error;
+  ok($MySQLiteObject::Objects_By_Key_Loaded{'name'}{'John'} != $loaded, "cache_expires_in uk 3 - $db_type");
+}
+
 BEGIN
 {
   #
@@ -892,6 +1077,85 @@ EOF
     Test::More::ok(MyInformixObject->meta->method_name_is_reserved('remember', 'MyInformixObject'), 'reserved method: remember');
     Test::More::ok(MyInformixObject->meta->method_name_is_reserved('forget', 'MyInformixObject'), 'reserved method: forget');
   }
+
+  #
+  # SQLite
+  #
+
+  eval 
+  {
+    $dbh = Rose::DB->new('sqlite_admin')->retain_dbh()
+      or die Rose::DB->error;
+  };
+
+  if(!$@ && $dbh)
+  {
+    our $HAVE_SQLITE = 1;
+
+    # Drop existing table and create schema, ignoring errors
+    {
+      local $dbh->{'RaiseError'} = 0;
+      local $dbh->{'PrintError'} = 0;
+      $dbh->do('DROP TABLE rose_db_object_test');
+    }
+
+    $dbh->do(<<"EOF");
+CREATE TABLE rose_db_object_test
+(
+  id             INTEGER PRIMARY KEY AUTOINCREMENT,
+  name           VARCHAR(32) NOT NULL,
+  flag           BOOLEAN NOT NULL,
+  flag2          BOOLEAN,
+  status         VARCHAR(32) DEFAULT 'active',
+  bits           VARCHAR(5) DEFAULT '00101' NOT NULL,
+  nums           VARCHAR(255),
+  start          DATE,
+  save           INT,
+  last_modified  TIMESTAMP,
+  date_created   TIMESTAMP,
+
+  UNIQUE(name)
+)
+EOF
+
+    $dbh->disconnect;
+
+    # Create test subclass
+
+    package MySQLiteObject;
+
+    our @ISA = qw(Rose::DB::Object::Cached);
+
+    sub init_db { Rose::DB->new('sqlite') }
+
+    MySQLiteObject->meta->table('rose_db_object_test');
+
+    MySQLiteObject->meta->columns
+    (
+      'name',
+      id       => { primary_key => 1 },
+      flag     => { type => 'boolean', default => 1 },
+      flag2    => { type => 'boolean' },
+      status   => { default => 'active' },
+      start    => { type => 'date', default => '12/24/1980' },
+      save     => { type => 'scalar' },
+      nums     => { type => 'array' },
+      bits     => { type => 'bitfield', bits => 5, default => 101 },
+      last_modified => { type => 'timestamp' },
+      date_created  => { type => 'timestamp' },
+    );
+
+    eval { MySQLiteObject->meta->initialize };
+    Test::More::ok($@, 'meta->initialize() reserved method');
+
+    MySQLiteObject->meta->add_unique_key('name');
+
+    MySQLiteObject->meta->alias_column(save => 'save_col');
+    MySQLiteObject->meta->initialize(preserve_existing => 1);
+
+    Test::More::ok(MySQLiteObject->meta->method_name_is_reserved('remember', 'MySQLiteObject'), 'reserved method: remember');
+    Test::More::ok(MySQLiteObject->meta->method_name_is_reserved('forget', 'MySQLiteObject'), 'reserved method: forget');
+  }
 }
 
 END
@@ -926,6 +1190,17 @@ END
   {
     # Informix
     my $dbh = Rose::DB->new('informix_admin')->retain_dbh()
+      or die Rose::DB->error;
+
+    $dbh->do('DROP TABLE rose_db_object_test');
+
+    $dbh->disconnect;
+  }
+
+  if($HAVE_SQLITE)
+  {
+    # SQLite
+    my $dbh = Rose::DB->new('sqlite_admin')->retain_dbh()
       or die Rose::DB->error;
 
     $dbh->do('DROP TABLE rose_db_object_test');
